Removed unnecessary DataParallel module manipulation. Added test case for get_original_model
